opae-tools: update hssi setup.py
  - move hssicommon.py, hssistats.py, hssiloopback.py, hssimac.py files move from top folder to ethernet.
  - remove pybind11 include in CMakelist.txt and setup.py.
  - remove build extensions from setup.py.
  - update pcie address command line input default value None.
