+++ tools/extra/hssi/CMakeLists.txt
@@ -24,16 +24,14 @@
 ## ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,  EVEN IF ADVISED OF THE
 ## POSSIBILITY OF SUCH DAMAGE.
 
-
 set(OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/timestamp)
 file(GLOB_RECURSE PKG_FILES ${CMAKE_CURRENT_SOURCE_DIR}/*.py)
 
-set(INCLUDE_DIRS "${OPAE_INCLUDE_PATH}:${pybind11_ROOT}/include")
 set(LINK_DIRS "${LIBRARY_OUTPUT_PATH}")
 
 add_custom_command(
     OUTPUT ${OUTPUT}
-    COMMAND ${PYTHON_EXECUTABLE} setup.py build_ext -I ${INCLUDE_DIRS} -L ${LINK_DIRS}
+    COMMAND ${PYTHON_EXECUTABLE} setup.py build_ext -L ${LINK_DIRS}
     COMMAND ${PYTHON_EXECUTABLE} setup.py build
     COMMAND ${CMAKE_COMMAND} -E touch ${OUTPUT}
     WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
